+++ Sources/PointFree/Account/Account.swift
@@ -106,1068 +106,3 @@ private func fetchAccountData<I>(
     }
     .sequential
 }
-
-private func accountView(_ data: AccountData) -> [Node] {
-  let content: [Node] =
-    [titleRowView]
-      + profileRowView(data)
-      + privateRssFeed(data)
-      + subscriptionOverview(data)
-      + creditsView(data)
-      + [logoutView]
-
-  return [
-    gridRow([
-      gridColumn(sizes: [.mobile: 12, .desktop: 8], [style(margin(leftRight: .auto))], [
-        div(
-          [`class`([Class.padding([.mobile: [.all: 3], .desktop: [.all: 4]])])],
-          content
-        )
-        ])
-      ])
-  ]
-}
-
-private func creditsView(_ data: AccountData) -> [Node] {
-  guard !data.subscriberState.isActive else { return [] }
-  guard data.currentUser.episodeCreditCount > 0 || data.episodeCredits.count > 0 else { return [] }
-
-  return [
-    gridRow([`class`([Class.padding([.mobile: [.bottom: 4]])])], [
-      gridColumn(sizes: [.mobile: 12], [
-        div(
-          [
-            h2([`class`([Class.pf.type.responsiveTitle4])], ["Episode Credits"]),
-            p([
-              "Episode credits allow you to see subscriber-only episodes before commiting to a full ",
-              .text("subscription. You currently have \(pluralizedCredits(count: data.currentUser.episodeCreditCount)) "),
-              "remaining."
-              ])
-            ]
-            + subscribeCallout(data.subscriberState)
-            + episodeCreditsView(data.episodeCredits)
-        )
-        ])
-      ])
-  ]
-}
-
-private func subscribeCallout(_ subscriberState: SubscriberState) -> [Node] {
-
-  guard !subscriberState.isActiveSubscriber else { return [] }
-
-  return [
-    p([
-      "To get all past and future episodes, ",
-      a([`class`([Class.pf.colors.link.purple]), href(path(to: .pricingLanding))], ["become"]),
-      " a subscriber today!"
-      ])
-  ]
-}
-
-private func pluralizedCredits(count: Int) -> String {
-  return count == 1
-    ? "1 credit"
-    : "\(count) credits"
-}
-
-private func episodeCreditsView(_ credits: [EpisodeCredit]) -> [Node] {
-  guard credits.count > 0 else { return [] }
-
-  return [
-    h3(
-      [
-        `class`(
-          [
-            Class.pf.type.responsiveTitle5,
-            Class.padding([.mobile: [.top: 2]])
-          ]
-        ),
-      ],
-      ["Chosen episodes"]
-    ),
-
-    ul(
-      episodes(from: credits)
-        .reversed()
-        .map { ep in li([episodeLinkView(ep)]) }
-    )
-  ]
-}
-
-private func episodeLinkView(_ episode: Episode) -> Node {
-  return a(
-    [
-      href(path(to: .episode(.left(episode.slug)))),
-      `class`(
-        [
-          Class.pf.colors.link.purple
-        ]
-      )
-    ],
-    [.text("#\(episode.sequence): \(episode.title)")]
-  )
-}
-
-private func episodes(from credits: [EpisodeCredit]) -> [Episode] {
-  return Current.episodes()
-    .filter { ep in credits.contains(where: { $0.episodeSequence == ep.sequence }) }
-}
-
-private func episode(atSequence sequence: Int) -> Episode? {
-  return Current.episodes()
-    .first(where: { $0.sequence == sequence })
-}
-
-private let titleRowView = gridRow([`class`([Class.padding([.mobile: [.bottom: 2]])])], [
-  gridColumn(sizes: [.mobile: 12], [
-    div([
-      h1([`class`([Class.pf.type.responsiveTitle2])], ["Account"])
-      ])
-    ])
-  ])
-
-private func profileRowView(_ data: AccountData) -> [Node] {
-
-  let nameFields = [
-    label([`class`([labelClass])], ["Name"]),
-    input([
-      `class`([blockInputClass]),
-      name(ProfileData.CodingKeys.name.stringValue),
-      type(.text),
-      value(data.currentUser.name ?? ""),
-      ])
-  ]
-
-  let emailFields = [
-    label([`class`([labelClass])], ["Email"]),
-    input([
-      `class`([blockInputClass]),
-      name(ProfileData.CodingKeys.email.stringValue),
-      type(.email),
-      value(data.currentUser.email.rawValue)
-      ])
-  ]
-
-  let showExtraInvoiceInfo = data.isSubscriptionOwner
-    && !data.subscriberState.isEnterpriseSubscriber
-  let extraInvoiceInfoFields = !showExtraInvoiceInfo ? [] : [
-    label([`class`([labelClass])], ["Extra Invoice Info"]),
-    textarea(
-      [
-        placeholder("Company name, billing address, VAT number, ..."),
-        `class`([blockInputClass]),
-        name(ProfileData.CodingKeys.extraInvoiceInfo.stringValue),
-      ],
-      data.stripeSubscription?.customer.right?.extraInvoiceInfo ?? ""
-    )
-  ]
-
-  let submit = [
-    input([
-      type(.submit),
-      `class`([Class.pf.components.button(color: .purple), Class.margin([.mobile: [.top: 3]])]),
-      value("Update profile")
-      ])
-  ]
-
-  let formContent: [Node] =
-    nameFields
-      + emailFields
-      + extraInvoiceInfoFields
-      + emailSettingCheckboxes(data.emailSettings, data.subscriberState)
-      + submit
-
-  return [
-    gridRow([`class`([Class.padding([.mobile: [.bottom: 4]])])], [
-      gridColumn(sizes: [.mobile: 12], [
-        div([
-          h2([`class`([Class.pf.type.responsiveTitle4])], ["Profile"]),
-
-          form(
-            [action(path(to: .account(.update(nil)))), method(.post)],
-            formContent
-          )
-          ])
-        ])
-      ])
-  ]
-}
-
-private func emailSettingCheckboxes(_ currentEmailSettings: [EmailSetting], _ subscriberState: SubscriberState) -> [Node] {
-  let newsletters = subscriberState.isNonSubscriber
-    ? EmailSetting.Newsletter.allNewsletters
-    : EmailSetting.Newsletter.subscriberNewsletters
-
-  return [
-    // TODO: hide `welcomeEmails` for subscribers?
-    p(["Receive email for:"]),
-    p([`class`([Class.padding([.mobile: [.left: 1]])])], newsletters.map { newsletter in
-      label([`class`([Class.display.block])], [
-        input(
-          [
-            type(.checkbox),
-            name("emailSettings[\(newsletter.rawValue)]"),
-            checked(currentEmailSettings.contains(where: ^\.newsletter == newsletter)),
-            `class`([Class.margin([.mobile: [.right: 1]])])
-          ]
-        ),
-        .text(newsletterDescription(newsletter))
-        ])
-    })
-  ]
-}
-
-private func newsletterDescription(_ type: EmailSetting.Newsletter) -> String {
-  switch type {
-  case .announcements:
-    return "New announcements (very infrequently)"
-  case .newBlogPost:
-    return "New blog posts on Point-Free Pointers (about every two weeks)"
-  case .newEpisode:
-    return "New episode is available (about once a week)"
-  case .welcomeEmails:
-    return "A short series of emails introducing Point-Free"
-  }
-}
-
-private func subscriptionOverview(_ data: AccountData) -> [Node] {
-
-  if data.isSubscriptionOwner {
-    return subscriptionOwnerOverview(data)
-  } else if data.stripeSubscription != nil {
-    return subscriptionTeammateOverview(data)
-  } else {
-    return []
-  }
-}
-
-private func privateRssFeed(_ data: AccountData) -> [Node] {
-  guard data.subscriberState.isActiveSubscriber else { return [] }
-  let user = data.currentUser
-  let encryptedUserId = Encrypted(user.id.rawValue.uuidString, with: Current.envVars.appSecret)
-  let encryptedRssSalt = Encrypted(user.rssSalt.rawValue.uuidString, with: Current.envVars.appSecret)
-  let rssUrl = zip(encryptedUserId, encryptedRssSalt)
-    .map { url(to: .account(.rss(userId: $0, rssSalt: $1))) }
-  let rssLink = rssUrl
-    .map { rssUrl in
-      [
-        p([
-          a(
-            [`class`([Class.pf.type.underlineLink]), href(rssUrl)],
-            [text(String(rssUrl.prefix(40)) + "...")]
-          )
-          ]
-        )
-        ]
-        + rssTerms(stripeSubscription: data.stripeSubscription)
-    }
-    ?? []
-
-  return [
-    gridRow([
-      gridColumn(
-        sizes: [.desktop: 10, .mobile: 12],
-        [style(margin(leftRight: .auto))],
-        [
-          div(
-            [
-              `class`(
-                [
-                  Class.margin([.mobile: [.bottom: 4]]),
-                  Class.padding([.mobile: [.all: 3]]),
-                  Class.pf.colors.bg.gray900
-                ]
-              )
-            ],
-            [
-              h4(
-                [
-                  `class`(
-                    [
-                      Class.pf.type.responsiveTitle4,
-                      Class.padding([.mobile: [.bottom: 2]])
-                    ]
-                  )
-                ],
-                ["Private RSS Feed"]
-              ),
-              p(["""
-Thanks for subscribing to Point-Free and supporting our efforts! We'd like to offer you an alternate way
-to consume our videos: an RSS feed that can be used with podcast apps!
-"""]),
-              p([
-                "The link below should work with most podcast apps out there today (please ",
-                a(
-                  [
-                    `class`([Class.pf.type.underlineLink]),
-                    href("mailto:support@pointfree.co")
-                  ],
-                  ["email us"]
-                ),
-                " if it doesn't). It is also tied directly to your Point-Free account and regularly ",
-                " monitored, so please do not share with others."
-                ])
-              ]
-              + rssLink
-          )
-        ]
-      )
-      ])
-  ]
-}
-
-private func rssTerms(stripeSubscription: Stripe.Subscription?) -> [Node] {
-  return stripeSubscription?.plan.interval == .some(.month)
-    ? [
-      p([
-        `class`([Class.padding([.mobile: [.all: 2]]), Class.margin([.mobile: [.top: 2]])]),
-        style(backgroundColor(.rgb(0xff, 0xff, 0xdd)))
-        ],
-        [
-          "Because you are on a monthly subscription plan, you get access to the last ",
-          .text("\(nonYearlyMaxRssItems)"),
-          " episodes in your RSS feed (don't worry, you can watch every past episode directly on this site).",
-          " To access all episodes from the RSS feed, please consider upgrading to a yearly subscription."
-        ]
-      )
-      ]
-    : []
-}
-
-private func subscriptionOwnerOverview(_ data: AccountData) -> [Node] {
-  guard let subscription = data.stripeSubscription else { return [] }
-
-  var content: [Node] = []
-  if data.subscriberState.isEnterpriseSubscriber {
-    content.append(contentsOf: enterpriseSubscriptionOverview(data))
-  } else {
-    content.append(contentsOf: subscriptionPlanRows(subscription, data.upcomingInvoice))
-    content.append(contentsOf: subscriptionTeamRow(data))
-    content.append(contentsOf: subscriptionInvitesRowView(data.teamInvites))
-    content.append(contentsOf: subscriptionInviteMoreRowView(data))
-    content.append(contentsOf: addTeammateToSubscriptionRow(data))
-    content.append(contentsOf: subscriptionPaymentInfoView(subscription))
-  }
-
-  return [
-    gridRow([`class`([Class.padding([.mobile: [.bottom: 4]])])], [
-      gridColumn(sizes: [.mobile: 12], [
-        div([
-          h2([`class`([Class.pf.type.responsiveTitle4])], ["Subscription overview"]),
-          gridColumn(sizes: [.mobile: 12], content)
-          ])
-        ])
-      ])
-  ]
-}
-
-private func enterpriseSubscriptionOverview(_ data: AccountData) -> [Node] {
-  guard let subscription = data.stripeSubscription else { return [] }
-  guard
-    case let .owner(_, _, .some(enterpriseAccount)) = data.subscriberState
-    else { return [] }
-
-  let planRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Plan"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([.text(planName(for: subscription))])
-            ])
-          ])
-        ])
-      ])
-    ])
-
-  let statusRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Status"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([.text(status(for: subscription))])
-            ])
-          ])
-        ])
-      ])
-    ])
-
-  let shareUrl = url(to: .enterprise(.landing(enterpriseAccount.domain)))
-  let shareRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Invite Link"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([
-              a(
-                [`class`([Class.pf.colors.link.purple]), href(shareUrl)],
-                [.text(shareUrl)]
-              )
-              ])
-            ])
-          ])
-        ])
-      ])
-    ])
-
-  let contactUsRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Contact Support"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([
-              a(
-                [`class`([Class.pf.colors.link.purple]), mailto("support@pointfree.co")],
-                ["support@pointfree.co"]
-              )
-            ])
-          ])
-        ])
-      ])
-    ])
-  ])
-
-  return [
-    div(
-      [`class`([Class.padding([.mobile: [.top: 1, .bottom: 3]])])],
-      [planRow, statusRow, shareRow, contactUsRow].compactMap(id)
-    )
-  ]
-}
-
-private func subscriptionTeammateOverview(_ data: AccountData) -> [Node] {
-  guard data.stripeSubscription != nil else { return [] }
-
-  var enterpriseShareLink: Node?
-  if case let .teammate(_, .some(enterpriseAccount)) = data.subscriberState {
-    let shareUrl = url(to: .enterprise(.landing(enterpriseAccount.domain)))
-    enterpriseShareLink = p([
-      "Share Point-Free with your co-workers by sending them this link: ",
-      a([href(shareUrl)], [.text(shareUrl)])
-      ])
-  }
-
-  return [
-    gridRow([`class`([Class.padding([.mobile: [.bottom: 4]])])], [
-      gridColumn(sizes: [.mobile: 12], [
-        div([
-          h2([`class`([Class.pf.type.responsiveTitle4])], ["Subscription overview"]),
-
-          p([
-            "You are currently on a team subscription. Contact ",
-            a(
-              [
-                mailto(data.subscriptionOwner?.email.rawValue ?? ""),
-                `class`([Class.pf.colors.link.purple])
-              ],
-              [
-                .text(data.subscriptionOwner?.name ?? "the owner")
-              ]
-            ),
-            " for more information.",
-            ]),
-
-          enterpriseShareLink,
-
-          form([action(path(to: .team(.leave))), method(.post)], [
-            input(
-              [
-                `class`([Class.pf.components.button(color: .red, size: .small)]),
-                type(.submit),
-                value("Leave this team")
-              ]
-            )
-            ]),
-          ].compactMap { $0 })
-        ])
-      ])
-  ]
-}
-
-private func planName(for subscription: Stripe.Subscription) -> String {
-  return subscription.quantity > 1
-    ? subscription.plan.nickname + " (×" + String(subscription.quantity) + ")"
-    : subscription.plan.nickname
-}
-
-public func status(for subscription: Stripe.Subscription) -> String {
-  switch subscription.status {
-  case .active:
-    let currentPeriodEndString = subscription.cancelAtPeriodEnd
-      ? " through " + dateFormatter.string(from: subscription.currentPeriodEnd)
-      : ""
-    return "Active" + currentPeriodEndString
-  case .canceled:
-    return "Canceled"
-  case .pastDue:
-    return "Past due"
-  case .unpaid:
-    return "Unpaid"
-  case .trialing:
-    return "In trial"
-  }
-}
-
-public func nextBilling(for subscription: Stripe.Subscription, upcomingInvoice: Stripe.Invoice?) -> String {
-  switch (subscription.status, upcomingInvoice) {
-  case let (.active, .some(invoice)):
-    return format(cents: invoice.amountDue)
-      + " on "
-      + dateFormatter.string(from: invoice.periodEnd)
-  case (.canceled, _):
-    return subscription.currentPeriodEnd > Current.date()
-      ? "Cancels " + dateFormatter.string(from: subscription.currentPeriodEnd)
-      : "Canceled"
-  case (.active, .none), (.pastDue, _), (.unpaid, _), (.trialing, _):
-    return "" // FIXME
-  }
-}
-
-private func subscriptionPlanRows(_ subscription: Stripe.Subscription, _ upcomingInvoice: Stripe.Invoice?) -> [Node] {
-
-  let planRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Plan"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([.text(planName(for: subscription))])
-            ])
-          ]),
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]]), Class.grid.end(.desktop)])], [
-            p([mainAction(for: subscription)])
-            ])
-          ])
-        ])
-      ])
-    ])
-
-  let statusRow = gridRow([
-    gridColumn(sizes: [.mobile: 3], [
-      p([div(["Status"])])
-      ]),
-    gridColumn(sizes: [.mobile: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([.text(status(for: subscription))])
-            ])
-          ]),
-        gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]]), Class.grid.end(.desktop)])], [
-            p(
-              subscription.isCancellable
-                ? [cancelAction(for: subscription)]
-                : []
-            )
-            ])
-          ])
-        ])
-      ])
-    ])
-
-  let nextBillingRow = subscription.cancelAtPeriodEnd || subscription.status == .canceled
-    ? nil
-    : gridRow([
-      gridColumn(sizes: [.mobile: 3], [
-        p([div(["Next billing"])])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-          p([.text(nextBilling(for: subscription, upcomingInvoice: upcomingInvoice))])
-          ])
-        ])
-      ])
-
-  let discountRow = subscription.discount.map { discount in
-    gridRow([
-      gridColumn(sizes: [.mobile: 3], [
-        p([div(["Discount"])])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-          p([.text(discountDescription(for: discount))])
-          ])
-        ])
-      ])
-  }
-
-  return [
-    div(
-      [`class`([Class.padding([.mobile: [.top: 1, .bottom: 3]])])],
-      [planRow, statusRow] + [nextBillingRow, discountRow].compactMap(id)
-    )
-  ]
-}
-
-private func discountDescription(for discount: Stripe.Discount) -> String {
-  return "\(discount.coupon.name ?? discount.coupon.id.rawValue): \(discount.coupon.formattedDescription)"
-}
-
-private func cancelAction(for subscription: Stripe.Subscription) -> Node {
-  return form(
-    [
-      action(path(to: .account(.subscription(.cancel)))),
-      method(.post),
-      onsubmit(unsafe: """
-if (!confirm("Cancel your subscription? You will lose access to Point-Free at the end of the current billing period. Should you change your mind, you can reactivate your subscription at any time before this period ends.")) {
-  return false
-}
-"""),
-    ],
-    [
-      button(
-        [`class`([Class.pf.components.button(color: .black, size: .small, style: .underline)])],
-        ["Cancel"]
-      )
-    ]
-  )
-}
-
-private func mainAction(for subscription: Stripe.Subscription) -> Node {
-  if subscription.isCanceling {
-    return form(
-      [action(path(to: .account(.subscription(.reactivate)))), method(.post)],
-      [
-        button(
-          [`class`([Class.pf.components.button(color: .purple, size: .small)])],
-          ["Reactivate"]
-        )
-      ]
-    )
-  } else if subscription.status == .canceled {
-    return a(
-      [
-        `class`([Class.pf.components.button(color: .purple, size: .small)]),
-        href(path(to: .pricingLanding))
-      ],
-      ["Resubscribe"]
-    )
-  } else {
-    switch subscription.plan.interval {
-    case .month:
-      let discount = subscription.discount?.coupon.discount ?? { $0 }
-      let amount = discount(subscription.quantity == 1 ? 168_00 : 144_00)
-        .map { $0 * subscription.quantity }
-      let formattedAmount = currencyFormatter.string(from: NSNumber(value: Double(amount.rawValue) / 100))
-      return form(
-        [
-          action(path(to: .account(.subscription(.change(.update(nil)))))),
-          method(.post),
-          onsubmit(unsafe: """
-if (!confirm("Upgrade to yearly billing? You will be charged \(formattedAmount ?? "") immediately with a prorated refund for the time remaining in your billing period.")) {
-  return false
-}
-"""),
-        ],
-        [
-          input([
-            name("billing"),
-            type(.hidden),
-            value("yearly"),
-          ]),
-          input([
-            name("quantity"),
-            type(.hidden),
-            value(subscription.quantity),
-          ]),
-          button(
-            [`class`([Class.pf.components.button(color: .purple, size: .small)])],
-            ["Upgrade to yearly billing"]
-          )
-        ]
-      )
-    case .year:
-      let discount = subscription.discount?.coupon.discount ?? { $0 }
-      let amount = discount(subscription.quantity == 1 ? 18_00 : 16_00)
-        .map { $0 * subscription.quantity }
-      let formattedAmount = currencyFormatter.string(from: NSNumber(value: Double(amount.rawValue) / 100))
-      return form(
-        [
-          action(path(to: .account(.subscription(.change(.update(nil)))))),
-          method(.post),
-          onsubmit(unsafe: """
-if (!confirm("Switch to monthly billing? You will be charged \(formattedAmount ?? "") on a monthly basis at the end of your current billing period.")) {
-  return false
-}
-"""),
-        ],
-        [
-          input([
-            name("billing"),
-            type(.hidden),
-            value("monthly"),
-          ]),
-          input([
-            name("quantity"),
-            type(.hidden),
-            value(subscription.quantity),
-          ]),
-          button(
-            [`class`([Class.pf.components.button(color: .purple, size: .small)])],
-            ["Switch to monthly billing"]
-          )
-        ]
-      )
-    }
-  }
-}
-
-private func subscriptionTeamRow(_ data: AccountData) -> [Node] {
-  guard
-    !data.teammates.isEmpty,
-    data.isTeamSubscription,
-    !data.subscriberState.isEnterpriseSubscriber
-    else { return [] }
-
-  return [
-    gridRow([`class`([subscriptionInfoRowClass])], [
-      gridColumn(sizes: [.mobile: 3], [
-        div([
-          p(["Team"])
-          ])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])],
-            [p(["Your current team:"])]
-              + data.teammates.map { teammateRowView(data.currentUser, $0) }
-        )
-        ])
-      ])
-  ]
-}
-
-private func teammateRowView(_ currentUser: User, _ teammate: User) -> Node {
-
-  let teammateLabel = currentUser.id == teammate.id
-    ? "\(teammate.displayName) (you)"
-    : teammate.name.map { "\($0) (\(teammate.email))" } ?? teammate.email.rawValue
-
-  return gridRow([
-    gridColumn(sizes: [.mobile: 8], [p([.text(teammateLabel)])]),
-    gridColumn(sizes: [.mobile: 4], [`class`([Class.grid.end(.desktop)])], [
-      form([action(path(to: .team(.remove(teammate.id)))), method(.post)], [
-        p([input([type(.submit), `class`([Class.pf.components.button(color: .purple, size: .small)]), value("Remove")])])
-        ]),
-      ])
-    ])
-}
-
-private func subscriptionInvitesRowView(_ invites: [TeamInvite]) -> [Node] {
-  guard !invites.isEmpty else { return [] }
-
-  return [
-    gridRow([`class`([subscriptionInfoRowClass])], [
-      gridColumn(sizes: [.mobile: 3], [
-        div([
-          p(["Invites"])
-          ])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])],
-            [p(["These teammates have been invited, but have not yet accepted."])]
-              <> invites.map(inviteRowView))
-        ])
-      ])
-  ]
-}
-
-private func inviteRowView(_ invite: TeamInvite) -> Node {
-  return gridRow([
-    gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-      p([.text(invite.email.rawValue)])
-      ]),
-    gridColumn(sizes: [.mobile: 12, .desktop: 6], [`class`([Class.grid.end(.desktop)])], [
-      form([action(path(to: .invite(.resend(invite.id)))), method(.post), `class`([Class.display.inlineBlock])], [
-        p([input([type(.submit), `class`([Class.pf.components.button(color: .purple, size: .small)]), value("Resend")])
-          ])]),
-
-      form([action(path(to: .invite(.revoke(invite.id)))), method(.post), `class`([Class.display.inlineBlock, Class.padding([.mobile: [.left: 1], .desktop: [.left: 2]])])], [
-        p([input([type(.submit), `class`([Class.pf.components.button(color: .red, size: .small, style: .underline)]), value("Revoke")])
-          ])]),
-      ]),
-    ])
-}
-
-private func addTeammateToSubscriptionRow(_ data: AccountData) -> [Node] {
-  guard !data.subscriberState.isEnterpriseSubscriber else { return [] }
-  guard let subscription = data.stripeSubscription else { return [] }
-  guard subscription.isRenewing else { return [] }
-  let invitesRemaining = subscription.quantity - data.teamInvites.count - data.teammates.count
-  guard invitesRemaining == 0 else { return [] }
-  
-  let amount = subscription.plan.interval == .some(.year) ? Cents(rawValue: 144_00) : Cents(rawValue: 16_00)
-  let interval = subscription.plan.interval == .some(.year) ? "year" : "month"
-
-  return [
-    gridRow([`class`([subscriptionInfoRowClass])], [
-      gridColumn(sizes: [.mobile: 3], [
-        div([
-          p(["Add teammate"])
-          ])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-          form([
-            action(path(to: .invite(.addTeammate(nil)))), method(.post),
-            `class`([Class.flex.flex, Class.padding([.mobile: [.top: 1]])])
-            ], [
-              input([
-                `class`([smallInputClass, Class.align.middle, Class.size.width100pct]),
-                name("email"),
-                placeholder("blob@example.com"),
-                type(.email),
-                ]),
-              input([
-                type(.submit),
-                `class`([
-                  Class.pf.components.button(color: .purple, size: .small),
-                  Class.align.middle,
-                  Class.margin([.mobile: [.left: 1], .desktop: [.left: 2]])
-                  ]),
-                value("Add")])
-            ]),
-
-          p(
-            [
-              `class`([
-                Class.pf.type.body.small,
-                Class.pf.colors.fg.gray400,
-                Class.padding([.mobile: [.top: 1]])
-                ])
-            ],
-            [.text("""
-Add a teammate to your subscription for a discounted rate of $\(amount.rawValue / 100)
-per \(interval). Your first invoice will be prorated based on your current billing cycle.
-""")]
-          ),
-          ])
-        ])
-      ])
-  ]
-}
-
-private func subscriptionInviteMoreRowView(_ data: AccountData) -> [Node] {
-
-  guard !data.subscriberState.isEnterpriseSubscriber else { return [] }
-  guard let subscription = data.stripeSubscription else { return [] }
-  guard subscription.quantity > 1 else { return [] }
-  let invites = data.teamInvites
-  let teammates = data.teammates
-  let invitesRemaining = subscription.quantity - invites.count - teammates.count
-  guard invitesRemaining > 0 else { return [] }
-
-  return [
-    gridRow([`class`([subscriptionInfoRowClass])], [
-      gridColumn(sizes: [.mobile: 3], [
-        div([
-          p(["Invite more"])
-          ])
-        ]),
-      gridColumn(sizes: [.mobile: 9], [
-        div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-          p([inviteTeammatesDescription(invitesRemaining: invitesRemaining)]),
-
-          form([
-            action(path(to: .invite(.send(nil)))), method(.post),
-            `class`([Class.flex.flex])
-            ], [
-              input([
-                `class`([smallInputClass, Class.align.middle, Class.size.width100pct]),
-                name("email"),
-                placeholder("blob@example.com"),
-                type(.email),
-                ]),
-              input([
-                type(.submit),
-                `class`([
-                  Class.pf.components.button(color: .purple, size: .small),
-                  Class.align.middle,
-                  Class.margin([.mobile: [.left: 1], .desktop: [.left: 2]])
-                  ]),
-                value("Invite")])
-            ])
-          ])
-        ])
-      ])
-  ]
-}
-
-private func inviteTeammatesDescription(invitesRemaining: Int) -> Node {
-  let seats = invitesRemaining == 1 ? "1 open seat": "\(invitesRemaining) open seats"
-  return .text("You have \(seats) on your team. Invite a team member below:")
-}
-
-private func subscriptionPaymentInfoView(_ subscription: Stripe.Subscription) -> [Node] {
-  guard let card = subscription.customer.right?.sources.data.first?.left
-    else { return subscriptionInvoiceBillingInfoView }
-
-  return [
-    gridRow([`class`([subscriptionInfoRowClass])], [
-      gridColumn(sizes: [.mobile: 3], [
-        div([
-          p(["Payment"])
-          ])
-        ]),
-      gridColumn(sizes: [.desktop: 9], [
-        gridRow([
-          gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-            div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-              p([.text(card.brand.rawValue + " ending in " + String(card.last4))]),
-              p([.text("Expires " + String(card.expMonth) + "/" + String(card.expYear))]),
-              ])
-            ]),
-          gridColumn(sizes: [.mobile: 12, .desktop: 6], [
-            div([`class`([Class.padding([.mobile: [.leftRight: 1]]), Class.grid.end(.desktop)])], [
-              p([
-                a([
-                  `class`([Class.pf.components.button(color: .purple, size: .small)]),
-                  href(path(to: .account(.paymentInfo(.show)))),
-                  ],
-                  ["Update payment method"])
-                ]),
-              p([
-                a([
-                  `class`([Class.pf.components.button(color: .black, size: .small, style: .underline)]),
-                  href(path(to: .account(.invoices(.index)))),
-                  ],
-                  ["Payment history"])
-                ])
-              ])
-            ])
-          ])
-        ])
-      ])
-  ]
-}
-
-private let subscriptionInvoiceBillingInfoView: [Node] = [
-  gridRow([`class`([subscriptionInfoRowClass])], [
-    gridColumn(sizes: [.mobile: 3], [
-      div([
-        p(["Payment"])
-        ])
-      ]),
-    gridColumn(sizes: [.desktop: 9], [
-      gridRow([
-        gridColumn(sizes: [.mobile: 12, .desktop: 7], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]])])], [
-            p([
-              "You are enrolled in invoice billing. If you have any questions, please ",
-              a(
-                [
-                  `class`([Class.pf.type.underlineLink]),
-                  href("mailto:support@pointfree.co")
-                ],
-                ["contact us"]
-              ),
-              "."
-              ])
-            ])
-          ]),
-        gridColumn(sizes: [.mobile: 12, .desktop: 5], [
-          div([`class`([Class.padding([.mobile: [.leftRight: 1]]), Class.grid.end(.desktop)])], [
-            p([
-              a([
-                `class`([Class.pf.components.button(color: .black, size: .small, style: .underline)]),
-                href(path(to: .account(.invoices(.index)))),
-                ],
-                ["Payment history"])
-              ])
-            ])
-          ])
-        ])
-      ])
-    ])
-]
-
-public func format(cents: Cents<Int>) -> String {
-  let dollars = NSNumber(value: Double(cents.rawValue) / 100)
-  return currencyFormatter.string(from: dollars)
-    ?? NumberFormatter.localizedString(from: dollars, number: .currency)
-}
-
-private let logoutView = gridRow([
-  gridColumn(sizes: [.mobile: 12], [
-    a([`class`([Class.pf.components.button(color: .black)]), href(path(to: .logout))], ["Logout"])
-    ])
-  ])
-
-private let subscriptionInfoRowClass =
-  Class.border.top
-    | Class.pf.colors.border.gray800
-    | Class.padding([.mobile: [.top: 2, .bottom: 3]])
-
-let labelClass =
-  Class.h5
-    | Class.type.bold
-    | Class.display.block
-    | Class.margin([.mobile: [.bottom: 1]])
-
-let baseInputClass =
-  Class.type.fontFamily.inherit
-    | Class.pf.colors.fg.black
-    | ".border-box"
-    | Class.border.rounded.all
-    | Class.border.all
-    | Class.pf.colors.border.gray800
-
-let regularInputClass =
-  baseInputClass
-    | Class.size.height(rem: 3)
-    | Class.padding([.mobile: [.all: 1]])
-    | Class.margin([.mobile: [.bottom: 2]])
-
-let smallInputClass =
-  baseInputClass
-    | Class.size.height(rem: 2)
-    | Class.padding([.mobile: [.all: 1]])
-
-let blockInputClass =
-  regularInputClass
-    | Class.size.width100pct
-    | Class.display.block
-
-let blockSelectClass =
-  Class.display.block
-    | Class.margin([.mobile: [.bottom: 2]])
-    | Class.size.height(rem: 3)
-    | Class.size.width100pct
-    | Class.pf.colors.border.gray800
-    | Class.type.fontFamily.inherit
-
-private struct AccountData {
-  let currentUser: User
-  let emailSettings: [EmailSetting]
-  let episodeCredits: [EpisodeCredit]
-  let stripeSubscription: Stripe.Subscription?
-  let subscriberState: SubscriberState
-  let subscription: Models.Subscription?
-  let subscriptionOwner: User?
-  let teamInvites: [TeamInvite]
-  let teammates: [User]
-  let upcomingInvoice: Stripe.Invoice?
-
-  var isSubscriptionOwner: Bool {
-    return self.currentUser.id == self.subscriptionOwner?.id
-  }
-
-  var isTeamSubscription: Bool {
-    return (self.stripeSubscription?.quantity ?? 0) > 1
-  }
-}
